Refactor FXIOS-5323 [v116] Fix large tuple in main menu action helper (#14868)
